requirements.txt
# pip install -r requirements.txt
# Python 3.9+ required

# Core dependencies
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1

# Data & ML
pandas==2.1.3
numpy==1.26.3
scikit-learn==1.3.2
scipy==1.11.4

# NLP & Embeddings
transformers==4.39.3
torch==2.3.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
huggingface-hub==0.23.2

# Frontend
streamlit==1.40.0
requests==2.31.0
pdfplumber==0.10.3
altair==5.1.2

# Utils
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
tqdm==4.66.1

# Testing (optional)
# Install with: pip install -r requirements.txt -e ".[dev]"
# Or: pip install pytest==7.4.3 pytest-asyncio==0.21.1 pytest-cov==4.1.0
